the timer now starts after file select, also better placement

+++ src/game/game_init.c
@@ -66,7 +66,7 @@ u8 gIsConsole;
 #ifdef WIDE
 u8 gWidescreen;
 #endif
-u16 gSpeedrunTimer;
+u64 gSpeedrunTimer;
 
 u16 sCurrFBNum = 0;
 u16 frameBufferIndex = 0;
