+++ src/game/game_init.h
@@ -45,7 +45,7 @@ extern u8 gIsConsole;
 #ifdef WIDE
 extern u8 gWidescreen;
 #endif
-extern u16 gSpeedrunTimer;
+extern u64 gSpeedrunTimer;
 #ifdef EEP
 extern s8 gEepromProbe;
 #endif
